requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
google-generativeai>=0.1.0
pymongo>=4.3.0
faiss-cpu>=1.7.0
numpy>=1.24.0
Pillow>=9.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0